+++ test/integration/distributed/inspection/etcd/suite_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package property_test
+package etcd_test
 
 import (
 	"testing"
@@ -23,31 +23,37 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/apache/skywalking-banyandb/pkg/test"
 	"github.com/apache/skywalking-banyandb/pkg/test/setup"
-	integration_standalone "github.com/apache/skywalking-banyandb/test/integration/standalone"
-	"github.com/apache/skywalking-banyandb/test/integration/standalone/inspect"
+	"github.com/apache/skywalking-banyandb/test/integration/distributed/inspection"
 )
 
 func init() {
-	inspect.SetupFunc = func() inspect.SetupResult {
-		By("Starting standalone server with property mode")
-		tmpDir, tmpDirCleanup, tmpErr := test.NewSpace()
-		Expect(tmpErr).NotTo(HaveOccurred())
-		dfWriter := setup.NewDiscoveryFileWriter(tmpDir)
-		config := setup.PropertyClusterConfig(dfWriter)
-		addr, _, closeFn := setup.EmptyStandalone(config)
-		return inspect.SetupResult{
-			Addr: addr,
+	inspection.SetupFunc = func() inspection.SetupResult {
+		By("Starting etcd server")
+		ep, _, etcdCleanup := setup.StartEmbeddedEtcd()
+		config := setup.EtcdClusterConfig(ep)
+
+		By("Starting data node 0")
+		closeDataNode0 := setup.DataNode(config)
+		By("Starting data node 1")
+		closeDataNode1 := setup.DataNode(config)
+		By("Starting liaison node")
+		liaisonAddr, closerLiaisonNode := setup.LiaisonNode(config)
+
+		return inspection.SetupResult{
+			LiaisonAddr:  liaisonAddr,
+			EtcdEndpoint: ep,
 			StopFunc: func() {
-				closeFn()
-				tmpDirCleanup()
+				closerLiaisonNode()
+				closeDataNode0()
+				closeDataNode1()
+				etcdCleanup()
 			},
 		}
 	}
 }
 
-func TestPropertyInspect(t *testing.T) {
+func TestEtcdInspect(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Standalone Property Inspect Suite", Label(integration_standalone.Labels...))
+	RunSpecs(t, "Distributed Etcd Inspect Suite")
 }